+++ classes/Kohana/Cookie.php
@@ -12,33 +12,60 @@ class Kohana_Cookie {
 
 	/**
 	 * @var  string  Magic salt to add to the cookie
+	 * @deprecated see Cookie::configure()
 	 */
 	public static $salt = NULL;
 
 	/**
 	 * @var  integer  Number of seconds before the cookie expires
+	 * @deprecated see Cookie::configure()
 	 */
 	public static $expiration = 0;
 
 	/**
 	 * @var  string  Restrict the path that the cookie is available to
+	 * @deprecated see Cookie::configure()
 	 */
 	public static $path = '/';
 
 	/**
 	 * @var  string  Restrict the domain that the cookie is available to
+	 * @deprecated see Cookie::configure()
 	 */
 	public static $domain = NULL;
 
 	/**
 	 * @var  boolean  Only transmit cookies over secure connections
+	 * @deprecated see Cookie::configure()
 	 */
-	public static $secure = FALSE;
+	public static $secure = TRUE;
 
 	/**
 	 * @var  boolean  Only transmit cookies over HTTP, disabling Javascript access
+	 * @deprecated see Cookie::configure()
 	 */
-	public static $httponly = FALSE;
+	public static $httponly = TRUE;
+
+	public static function configure(array $options)
+	{
+		$values             = array_merge(
+			[
+				'salt'       => Cookie::$salt,
+				'expiration' => Cookie::$expiration,
+				'path'       => Cookie::$path,
+				'domain'     => Cookie::$domain,
+				'secure'     => Cookie::$secure,
+				'httponly'   => Cookie::$httponly
+			],
+			$options
+		);
+		Cookie::$salt       = $values['salt'];
+		Cookie::$expiration = $values['expiration'];
+		Cookie::$path       = $values['path'];
+		Cookie::$domain     = $values['domain'];
+		Cookie::$secure     = $values['secure'];
+		Cookie::$httponly   = $values['httponly'];
+	}
 
 	/**
 	 * Gets the value of a signed cookie. Cookies without signatures will not
@@ -102,8 +129,10 @@ class Kohana_Cookie {
 	 * @return  boolean
 	 * @uses    Cookie::salt
 	 */
-	public static function set($name, $value, $lifetime = NULL)
+	public static function set($name, $value, $lifetime = NULL, array $options = [])
 	{
+	        $options = static::mergeCookieOptions($options);
+
 		if ($lifetime === NULL)
 		{
 			// Use the default expiration
@@ -119,7 +148,33 @@ class Kohana_Cookie {
 		// Add the salt to the cookie value
 		$value = Cookie::salt($name, $value).'~'.$value;
 
-		return static::_setcookie($name, $value, $lifetime, Cookie::$path, Cookie::$domain, Cookie::$secure, Cookie::$httponly);
+	        return static::_setcookie(
+			$name,
+			$value,
+			$lifetime,
+			$options['path'],
+			$options['domain'],
+			$options['secure'],
+			$options['httponly']
+		);
+	}
+
+	protected static function mergeCookieOptions(array $options):array
+	{
+		$defaults = [
+			'path'       => Cookie::$path,
+			'domain'     => Cookie::$domain,
+			'secure'     => Cookie::$secure,
+			'httponly'   => Cookie::$httponly
+		];
+
+		if ($extras = array_diff_key($options, $defaults)) {
+			throw new \InvalidArgumentException(
+				'Cannot customise '.implode(', ', array_keys($extras)).' for individual cookies'
+			);
+		}
+
+		return array_merge($defaults, $options);
 	}
 
 	/**