+++ tests/kohana/CookieTest.php
@@ -36,6 +36,77 @@ class Kohana_CookieTest extends Unittest_TestCase
 		));
 	}
 
+
+	public function provider_configure()
+	{
+		return [
+			[
+				[],
+				[
+					'salt'       => 'foo',
+					'expiration' => 10,
+					'path'       => '/subdir',
+					'domain'     => 'foo.domain',
+					'secure'     => TRUE,
+					'httponly'   => TRUE
+				],
+				[
+					'salt' => 'foo',
+					'expiration' => 10,
+					'path'       => '/subdir',
+					'domain'     => 'foo.domain',
+					'secure'     => TRUE,
+					'httponly'   => TRUE
+				],
+			],
+			[
+				[
+					'Cookie::$salt'     => 'anything',
+					'Cookie::$path'     => '/path',
+					'Cookie::$domain'   => 'my.domain',
+					'Cookie::$secure'   => TRUE,
+					'Cookie::$httponly' => FALSE,
+				],
+				[
+					'expiration' => 30,
+					'secure'     => FALSE,
+					'httponly'   => TRUE
+				],
+				[
+					'salt' => 'anything',
+					'expiration' => 30,
+					'path'       => '/path',
+					'domain'     => 'my.domain',
+					'secure'     => FALSE,
+					'httponly'   => TRUE
+				],
+			]
+		];
+	}
+
+	/**
+	 * @dataProvider provider_configure
+	 */
+	public function test_configure_updates_options_merging_defaults(
+		$env,
+		$config,
+		$expect
+	) {
+		$this->setEnvironment($env);
+		Cookie::configure($config);
+		$this->assertSame(
+			$expect,
+			[
+				'salt'       => Cookie::$salt,
+				'expiration' => Cookie::$expiration,
+				'path'       => Cookie::$path,
+				'domain'     => Cookie::$domain,
+				'secure'     => Cookie::$secure,
+				'httponly'   => Cookie::$httponly
+			]
+		);
+	}
+
 	/**
 	 * Tests that cookies are set with the global path, domain, etc options.
 	 *
@@ -86,6 +157,55 @@ class Kohana_CookieTest extends Unittest_TestCase
 		$this->setEnvironment(array('Cookie::$expiration' => self::COOKIE_EXPIRATION));
 		Kohana_CookieTest_TestableCookie::set('foo', 'bar', $expiration);
 		$this->assertSetCookieWith(array('expire' => $expect_expiry));
+	}
+
+	public function test_set_can_take_custom_cookie_options()
+	{
+		Cookie::configure(
+			[
+				'salt'       => 'foo',
+				'expiration' => 10,
+				'path'       => '/subdir',
+				'domain'     => 'foo.domain',
+				'secure'     => TRUE,
+				'httponly'   => TRUE
+			]
+		);
+		Kohana_CookieTest_TestableCookie::set(
+			'bespoke',
+			'any-val',
+			0,
+			[
+				'path'     => '/my-path',
+				'domain'   => '*',
+				'secure'   => FALSE,
+				'httponly' => FALSE
+			]
+		);
+		$this->assertSetCookieWith(
+			[
+				'expire'   => 0,
+				'path'     => '/my-path',
+				'domain'   => '*',
+				'secure'   => FALSE,
+				'httponly' => FALSE
+			]
+		);
+	}
+
+	/**
+	 * @testWith ["salt"]
+	 *           ["expiration"]
+	 */
+	public function test_set_throws_on_unsupported_cookie_options($opt)
+	{
+		$this->expectException(\InvalidArgumentException::class);
+		Kohana_CookieTest_TestableCookie::set(
+			'any',
+			'thing',
+			0,
+			[$opt => 'anything']
+		);
 	}
 
 	/**